Rename Game module to replace old one
